+++ packages/util/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 (modification: no type change headlines) and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Features
+
+- Add `generateTronCreateAddress(rootTransactionId, nonce)` for java-tron-compatible internal CREATE address derivation while preserving Ethereum `generateAddress()` behavior
+
 ## 1.0.0
 
 ### Features
@@ -29,5 +35,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Rename package namespace from `@ethereumjs/util` to `@tvmjs/util`
 - Bump package version to `1.0.0`
 - Lock all dependency versions by removing `^` and `~` prefixes
-